// donnyadmin/src/types/estadistica.d.ts
export interface Estadistica {
  ndocumentos: number;
  nhojas: number;
  nbytes: number;
  ffecha: string;
  nvideos: number;
  nlaboral: number;
  nfamilia: number;
  nresoluciones: number;
  npenal: number;
  nidEstadistica: number;
  ncivil: number;
  nactas: number;
  faud: string | null;
  nidModulo: number;
  baud: string | null;
}